[Silver III] Title: 주유소, Time: 340 ms, Memory: 30412 KB -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -16,7 +16,7 @@ class Main{
 		for(int i=0; i<pri.length; i++) {
 			pri[i]=Integer.parseInt(st.nextToken());
 		}
-		int result = pri[0]*dis[0]; int price = pri[0]>pri[1]?pri[1]:pri[0];
+		long result = pri[0]*dis[0]; int price = pri[0]>pri[1]?pri[1]:pri[0];
 		for(int i=1; i<dis.length; i++) {
 			result+=price*dis[i];
 			if(pri[i]>pri[i+1]) {
